digiClass_backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg==3.1.12
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
python-decouple==3.8
phonenumbers==8.13.26